// old/features/account/types/types.ts
export interface AvatarOption {
    id: string
    url: string
    provider: string
    label: string
    icon?: string
    isExternal?: boolean
    fileName?: string
    size?: number
    uploadedAt?: string | Date
}